refactor: address PR #522 review — finish /import removal, fix publish wiring, relocate TokenUsage
Fresh multi-agent review surfaced two blockers + cleanups; all resolved.

Blocker 1 — incomplete /import removal (would break integration + e2e CI):
The deprecated route was deleted but its consumers were not. Repointed every
remaining caller to the canonical /pair/redeem:
- 3 integration test files still POSTed to /api/model-providers-oauth/import
  (model-providers-oauth.test.ts, the bearer-track suite, and a legacy-alias
  describe block in pair-redeem.test.ts). Repointed the route + bearer suites
  (renamed the bearer file ...-import-pairing-bearer → ...-pair-redeem-bearer)
  and deleted the deprecation-header legacy-alias block (asserted headers that
  no longer exist).
- 2 e2e specs (pairing-lifecycle, oauth-flow) repointed.
- Removed the now-dead /import auth-bypass branch in auth-pipeline.skipAuth.
- Swept stale /import references in service/schema doc comments.

Blocker 2 — published-package dependency wiring:
@appstrate/core is published via `npm publish`, which does not rewrite the
`workspace:` protocol. Changed core's @appstrate/afps-shared dep from
`workspace:*` to the range `^0.1.0` (matching the @appstrate/ui→core
convention) so external `npm install @appstrate/core` resolves. Documented the
publish-order constraint (afps-shared before core) in the publish workflow.

Cleanups:
- Relocated the canonical TokenUsage type from @appstrate/core/token-usage into
  the zero-dep leaf @appstrate/afps-shared/token-usage, so afps-runtime/db no
  longer import from core (honors the leaf-isolation rationale). core re-exports
  it for path stability; db deps afps-shared (^0.1.0) instead of core for this.
- run-info-tab.tsx uses the canonical TokenUsage instead of an inline copy.
- Removed a stale "legacy un-prefixed signatures accepted" comment in
  pending-client-cookie.ts (the fallback was removed earlier in this PR).

Verified: bun run check 22/22; OAuth route+bearer suites 13/0; afps-shared +
core + afps-runtime + runner-pi 1290/0; api unit 809/0.
