+++ apps/api/test/integration/routes/model-providers-oauth.test.ts
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /**
- * Route-level tests for `POST /api/model-providers-oauth/import`.
+ * Route-level tests for `POST /api/model-providers-oauth/pair/redeem`.
  *
- * The route is bearer-only (pairing token). The happy-path import flow + the
+ * The route is bearer-only (pairing token). The happy-path redeem flow + the
  * pairing/credentials-mutation contract is covered in
- * `model-providers-oauth-import-pairing-bearer.test.ts`; this file pins the
+ * `model-providers-oauth-pair-redeem-bearer.test.ts`; this file pins the
  * post-consume Zod refine and the bearer-required check.
  */
 
@@ -27,7 +27,7 @@ async function mintPairing(ctx: TestContext, providerId = "test-oauth"): Promise
   return body.token;
 }
 
-describe("POST /api/model-providers-oauth/import", () => {
+describe("POST /api/model-providers-oauth/pair/redeem", () => {
   let ctx: TestContext;
 
   beforeEach(async () => {
@@ -37,7 +37,7 @@ describe("POST /api/model-providers-oauth/import", () => {
 
   it("returns 400 when the body providerId does not match the pairing", async () => {
     const token = await mintPairing(ctx, "test-oauth");
-    const res = await app.request("/api/model-providers-oauth/import", {
+    const res = await app.request("/api/model-providers-oauth/pair/redeem", {
       method: "POST",
       headers: { "Content-Type": "application/json", Authorization: `Bearer ${token}` },
       body: JSON.stringify({
@@ -51,7 +51,7 @@ describe("POST /api/model-providers-oauth/import", () => {
   });
 
   it("returns 401 without authentication", async () => {
-    const res = await app.request("/api/model-providers-oauth/import", {
+    const res = await app.request("/api/model-providers-oauth/pair/redeem", {
       method: "POST",
       headers: { "Content-Type": "application/json" },
       body: JSON.stringify({
@@ -67,7 +67,7 @@ describe("POST /api/model-providers-oauth/import", () => {
   it("returns 401 when a session cookie is presented without a pairing-bearer", async () => {
     // The route no longer accepts session auth — cookie callers reach the
     // handler (auth-pipeline only bypasses on `Bearer appp_`) and 401 there.
-    const res = await app.request("/api/model-providers-oauth/import", {
+    const res = await app.request("/api/model-providers-oauth/pair/redeem", {
       method: "POST",
       headers: authHeaders(ctx, { "Content-Type": "application/json" }),
       body: JSON.stringify({
