+++ packages/afps-runtime/src/types/run-result.ts
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright 2026 Appstrate
 
-import type { TokenUsage } from "@appstrate/core/token-usage";
+import type { TokenUsage } from "@appstrate/afps-shared/token-usage";
 
 /**
  * Severity levels carried by `log.written` run events. Mirrored on the
@@ -83,8 +83,8 @@ export interface RunResult {
  * `appstrate.metric` event. Mirrors the platform's `runs.tokenUsage` JSONB
  * column shape so finalize can persist it directly without re-mapping.
  *
- * Canonical definition lives in `@appstrate/core/token-usage` (imported at the
- * top of this file); re-exported here so existing `@appstrate/afps-runtime`
+ * Canonical definition lives in `@appstrate/afps-shared/token-usage` (imported
+ * at the top of this file); re-exported here so existing `@appstrate/afps-runtime`
  * importers keep working.
  */
 export type { TokenUsage };
