+++ packages/afps-shared/package.json
@@ -40,7 +40,8 @@
     "./integrity": "./src/integrity.ts",
     "./credential-template": "./src/credential-template.ts",
     "./delivery-http": "./src/delivery-http.ts",
-    "./ssrf": "./src/ssrf.ts"
+    "./ssrf": "./src/ssrf.ts",
+    "./token-usage": "./src/token-usage.ts"
   },
   "dependencies": {
     "semver": "^7.7.1"
