+++ packages/core/package.json
@@ -81,7 +81,7 @@
   },
   "dependencies": {
     "@afps-spec/schema": "^0.1.0",
-    "@appstrate/afps-shared": "workspace:*",
+    "@appstrate/afps-shared": "^0.1.0",
     "ajv": "^8.20.0",
     "ajv-formats": "^3.0.1",
     "fflate": "^0.8.0",
